requirements.txt
azure-iot-device
CounterFit
counterfit-connection
counterfit-shims-grove
mockito
pytest
flake8